refactor: move Utilities unit to src/Utilities subdirectory
