+++ tests/Editor.test.cpp
@@ -24,7 +24,7 @@
 #include "Editor.hpp"
 
 #include "Cursor/Cursor.hpp"
-#include "Utilities.hpp"
+#include "Utilities/Utilities.hpp"
 #include <cstring>
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
